Controller adicionado
